Cleanup and overhaul for API v3
