+++ circle.yml
@@ -1,4 +1,3 @@
-
 machine:
   services:
     - docker
@@ -10,4 +9,4 @@ test:
   override:
     - tests/run.sh
   post:
-    - docker run -e "COVERALLS_REPO_TOKEN=$COVERALLS_REPO_TOKEN" -w /srv/mailman-api -v $PWD:/srv/mailman-api tracywebtech/mailman-api-test coveralls
+    - docker run -e "COVERALLS_REPO_TOKEN=$COVERALLS_REPO_TOKEN" -w /srv/mailman-api -v $PWD:/srv/mailman-api mailman-api-test coveralls
