+++ docs/api.rst
@@ -7,10 +7,7 @@ The body of all responses contains valid JSON objects. Unless otherwise noted, s
 
 Supported methods:
 
-API V1
-------
-
-list_lists
+List Lists
 ++++++++++
 Lists existing mailing lists on the server.
 
@@ -20,49 +17,51 @@ Lists existing mailing lists on the server.
 
     Returns a list of the mailing lists that exist on this server.
 
-subscribe
-+++++++++
-Adds a new subscriber to the list called `<listname>`
+    **Parameters**:
+        * `address` (optional): email address to search for in lists.
+
+Create List
++++++++++++
+Create a new list.
 
     **Method**: PUT
 
     **URI**: /<listname>
 
     **Parameters**:
+        * `admin`: email of list admin
+        * `password`: list admin password
+        * `subscribe_policy`: 1) Confirm; 2) Approval; 3)Confirm and approval.
+          Default is Confirm (1)
+        * `archive_private`: 0) Public; 1) Private. Default is Public (0)
 
-      * `address`: email address that is to be subscribed to the list.
-      * `fullname`: full name of the person being subscribed to the list.
-      * `digest`: if this equals `true`, the new subscriber will receive
-        digests instead of every mail sent to the list.
-
-unsubscribe
+Delete List
 +++++++++++
-Unsubscribe an email address from the mailing list.
+Delete a list.
 
     **Method**: DELETE
 
     **URI**: /<listname>
 
-    **Parameters**:
 
-      * `address`: email address that is to be unsubscribed from the list
-
-members
-+++++++
-Lists subscribers for the `listname` list.
+List Attributes
++++++++++++++++
+Get the list attributes.
 
     **Method**: GET
 
     **URI**: /<listname>
 
-    Returns an array of email addresses.
-sendmail
-++++++++
-Posts an email to the mailing list.
+    Returns a dictionary containing the basic attributes for a specific mailing
+    list that exist on this server.
+
+Send Mail
++++++++++
+Send an email to the list.
 
     **Method**: POST
 
-    **URI**: /<listname>/sendmail
+    **URI**: /<listname>
 
     **Parameters**:
 
@@ -73,42 +72,13 @@ Posts an email to the mailing list.
       * `in_reply_to` (optional): Message-ID of the message that is being
         replied to, if any.
 
-API V2
-------
-
-list_lists
-++++++++++
-Lists existing mailing lists on the server.
-
-    **Method**: GET
-
-    **URI**: /v2/
-
-    Returns a list of dictionaries the mailing lists and its public attributes that exist on this server.
-
-create_list
-+++++++++++
-Create a mail list.
-
-    **Method**: POST
-
-    **URI**: /v2/lists/<listname>
-
-    **Parameters**:
-
-      * `admin`: email of list admin
-      * `password`: list admin password
-      * `subscription_policy`: 1) Confirm; 2) Approval; 3)Confirm and approval.
-      Default is Confirm (1)
-      * `archive_privacy`: 0) Public; 1) Private. Default is Public (0)
-
-subscribe
+Subscribe
 +++++++++
 Adds a new subscriber to the list called `<listname>`
 
     **Method**: PUT
 
-    **URI**: /v2/subscribe/<listname>
+    **URI**: /<listname>/members
 
     **Parameters**:
 
@@ -117,44 +87,27 @@ Adds a new subscriber to the list called `<listname>`
       * `digest`: if this equals `true`, the new subscriber will receive
         digests instead of every mail sent to the list.
 
-unsubscribe
+Unsubscribe
 +++++++++++
 Unsubscribe an email address from the mailing list.
 
     **Method**: DELETE
 
-    **URI**: /v2/subscribe/<listname>
+    **URI**: /<listname>/members
 
     **Parameters**:
 
       * `address`: email address that is to be unsubscribed from the list
 
-members
+Members
 +++++++
 Lists subscribers for the `listname` list.
 
     **Method**: GET
 
-    **URI**: /v2/members/<listname>
-
-    Returns an array of email addresses.
-
-sendmail
-++++++++
-Posts an email to the mailing list.
-
-    **Method**: POST
-
-    **URI**: /v2/sendmail/<listname>
+    **URI**: /<listname>/members
 
     **Parameters**:
+        * `address` (optional): email address to search for in list.
 
-      * `name_from`: name of the poster
-      * `email_from`: email address of the poster
-      * `subject`: the subject of the message
-      * `body`: the body of the message.
-      * `in_reply_to` (optional): Message-ID of the message that is being
-        replied to, if any.
-
-.. automodule:: mailmanapi.api
-      :members:
+    Returns an array of email addresses.
