+++ mailmanapi/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '0.3rc4'
+__version__ = '0.1'
