+++ mailmanapi/members.py
@@ -1,4 +1,3 @@
-
 class Member(object):
     def __init__(self, fullname=None, address=None, digest=None):
         self.fullname = fullname
