+++ mailmanapi/utils.py
@@ -1,10 +1,6 @@
-
-
 import json
 import logging
-
 from time import strftime
-
 from bottle import HTTPResponse
 
 try:
@@ -12,13 +8,10 @@ try:
 except ImportError:
     logging.error('Could not import Mailman module')
 
-
 def parse_boolean(value):
     if value and value.lower() == 'true':
         return True
-
     return False
-
 
 def jsonify(obj, status=200):
     response = HTTPResponse(content_type='application/json')
@@ -26,17 +19,15 @@ def jsonify(obj, status=200):
     response.status = status
     return response
 
-
 def get_mailinglist(listname, lock=True):
     try:
         return MailList.MailList(listname, lock=lock)
     except Errors.MMUnknownListError:
         raise jsonify("Unknown Mailing List `{}`.".format(listname), 404)
 
-
 def get_timestamp():
     return strftime('%a, %d %b %Y %H:%M:%S %z (%Z)')
-    
+
 def remove_list(listname):
     #TODO
     #see /usr/lib/mailman/bin/rmlist
