+++ provision.sh
@@ -41,5 +41,3 @@ sudo cp init/systemd/mailman-api.service /usr/lib/systemd/system
 sudo ln -s /usr/lib/systemd/system/mailman-api.service /etc/systemd/system/multi-user.target.wants/mailman-api.service
 
 sudo systemctl start mailman-api
-
-
