+++ requirements.txt
@@ -1,3 +1,4 @@
+setuptools>=20.10
 Paste==1.7.5.1
 bottle==0.11.6
 bottle-beaker==0.1.0
