+++ tests/Dockerfile
@@ -1,12 +1,15 @@
-
 FROM ubuntu:14.04
 
-MAINTAINER devops@tracy.com.br
+MAINTAINER epoitras@linuxfoundation.org
 
 ENV DEBIAN_FRONTEND=noninteractive
 
 RUN apt-get update
 
-RUN apt-get install -y mailman python-pip python-nose git
+RUN apt-get install -y python-dev mailman python-pip python-nose git
+
+RUN pip install --upgrade pip
+
+RUN pip install --upgrade setuptools
 
 RUN pip install coverage flake8 coveralls
