+++ tests/__init__.py
@@ -1,4 +1,3 @@
-
 import sys
 import bottle
