+++ tests/run.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-
 if [ "$(uname)" == 'Darwin' ]; then
     eval "$(boot2docker shellinit)"
 fi
@@ -33,7 +32,7 @@ function run {
 
 DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 
-DOCKER_IMAGE="tracywebtech/mailman-api-test"
+DOCKER_IMAGE="mailman-api-test"
 
 DOCKER_CMD="docker run -w /srv/mailman-api -v $DIR/../:/srv/mailman-api $DOCKER_IMAGE"
