+++ tests/test_api.py
@@ -1,106 +1,101 @@
-
-
 from .utils import MailmanAPITestCase
 from Mailman import MailList, UserDesc, Defaults
 from time import strftime
 
-
-class TestAPIv2(MailmanAPITestCase):
-    api_version = 'API V2'
-    url = '/v2/'
+class TestAPI(MailmanAPITestCase):
+    url = '/'
     data = {'address': 'user@email.com'}
     list_name = 'list'
 
     def setUp(self):
-        super(TestAPIv2, self).setUp()
+        super(TestAPI, self).setUp()
         self.create_list(self.list_name)
 
     def tearDown(self):
-        super(TestAPIv2, self).tearDown()
+        super(TestAPI, self).tearDown()
         self.remove_list(self.list_name)
 
     def test_subscribe_no_moderation(self):
-        path = 'subscribe/'
+        path = '/members'
 
         self.change_list_attribute('subscribe_policy', 0)
-        resp = self.client.put(self.url + path + self.list_name,
+        resp = self.client.put(self.url + self.list_name + path,
                                self.data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(0, resp.json)
 
     def test_subscribe_confirm(self):
-        path = 'subscribe/'
+        path = '/members'
 
         self.change_list_attribute('subscribe_policy', 1)
-        resp = self.client.put(self.url + path + self.list_name,
+        resp = self.client.put(self.url + self.list_name + path,
                                self.data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(1, resp.json)
 
     def test_subscribe_approval(self):
-        path = 'subscribe/'
+        path = '/members'
 
         self.change_list_attribute('subscribe_policy', 2)
-        resp = self.client.put(self.url + path + self.list_name,
+        resp = self.client.put(self.url + self.list_name + path,
                                self.data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(2, resp.json)
 
     def test_subscribe_banned(self):
-        path = 'subscribe/'
+        path = '/members'
         mlist = MailList.MailList(self.list_name)
         mlist.ban_list.append(self.data['address'])
         mlist.Save()
         mlist.Unlock()
 
-        resp = self.client.put(self.url + path + self.list_name,
+        resp = self.client.put(self.url + self.list_name + path,
                                self.data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(4, resp.json)
 
     def test_subscribe_already_member(self):
-        path = 'subscribe/'
+        path = '/members'
         user_desc = UserDesc.UserDesc(self.data['address'], 'fullname', 1)
         mlist = MailList.MailList(self.list_name)
         mlist.AddMember(user_desc)
         mlist.Save()
         mlist.Unlock()
 
-        resp = self.client.put(self.url + path + self.list_name,
+        resp = self.client.put(self.url + self.list_name + path,
                                self.data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(3, resp.json)
 
     def test_subscribe_bad_email(self):
-        path = 'subscribe/'
+        path = '/members'
         data = {'address': 'user@emailcom'}
-        resp = self.client.put(self.url + path + self.list_name,
+        resp = self.client.put(self.url + self.list_name + path,
                                data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(5, resp.json)
 
     def test_unsubscribe(self):
-        path = 'subscribe/'
+        path = '/members'
         user_desc = UserDesc.UserDesc(self.data['address'], 'fullname', 1)
         mlist = MailList.MailList(self.list_name)
         mlist.AddMember(user_desc)
         mlist.Save()
         mlist.Unlock()
 
-        resp = self.client.delete(self.url + path + self.list_name,
+        resp = self.client.delete(self.url + self.list_name + path,
                                   self.data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(0, resp.json)
 
     def test_unsubscribe_not_member(self):
-        path = 'subscribe/'
-        resp = self.client.delete(self.url + path + self.list_name,
+        path = '/members'
+        resp = self.client.delete(self.url + self.list_name + path,
                                   self.data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(7, resp.json)
 
     def test_sendmail(self):
-        path = 'sendmail/'
         mlist = MailList.MailList(self.list_name)
         data = {}
         data['email_to'] = mlist.GetListEmail()
@@ -112,13 +107,12 @@ class TestAPIv2(MailmanAPITestCase):
         data['subject'] = 'subject test'
         data['body'] = 'body test'
 
-        resp = self.client.post(self.url + path + self.list_name,
+        resp = self.client.post(self.url + self.list_name,
                                 data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(0, resp.json)
 
     def test_sendmail_with_reply(self):
-        path = 'sendmail/'
         mlist = MailList.MailList(self.list_name)
         data = {}
         data['email_to'] = mlist.GetListEmail()
@@ -131,21 +125,20 @@ class TestAPIv2(MailmanAPITestCase):
         data['body'] = 'body test'
         data['in_reply_to'] = 1
 
-        resp = self.client.post(self.url + path + self.list_name,
+        resp = self.client.post(self.url + self.list_name,
                                 data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(0, resp.json)
 
     def test_sendmail_missing_information(self):
-        path = 'sendmail/'
         data = {}
-        resp = self.client.post(self.url + path + self.list_name,
+        resp = self.client.post(self.url + self.list_name,
                                 data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(8, resp.json)
 
     def test_sendmail_unknown_list(self):
-        path = 'sendmail/'
+        path = ''
         data = {}
 
         resp = self.client.post(self.url + path + 'unknown_list',
@@ -154,12 +147,11 @@ class TestAPIv2(MailmanAPITestCase):
         self.assertEqual(12, resp.json)
 
     def test_mailman_site_list_not_listed_among_lists(self):
-        path = 'lists/'
         mailman_site_list = Defaults.MAILMAN_SITE_LIST
 
         self.create_list(mailman_site_list)
 
-        resp = self.client.get(self.url + path, expect_errors=True)
+        resp = self.client.get(self.url, expect_errors=True)
 
         self.assertEqual(resp.status_code, 200)
         self.assertIsInstance(resp.json, list)
@@ -169,9 +161,7 @@ class TestAPIv2(MailmanAPITestCase):
             self.assertNotEqual(mlist.get("listname"), mailman_site_list)
 
     def test_list_lists(self):
-        path = 'lists/'
-
-        resp = self.client.get(self.url + path, expect_errors=True)
+        resp = self.client.get(self.url, expect_errors=True)
         total_lists = len(resp.json)
         found = False
 
@@ -188,32 +178,32 @@ class TestAPIv2(MailmanAPITestCase):
 
     def test_create_list(self):
         new_list = 'new_list'
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456'}
 
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(0, resp.json)
         self.remove_list(new_list)
 
     def test_create_list_already_exists(self):
         new_list = self.list_name
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456'}
 
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(13, resp.json)
 
     def test_create_private_list(self):
         new_list = "new_list"
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456',
                 'archive_private': 1}
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         mlist = MailList.MailList(new_list)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(int(mlist.archive_private), 1)
@@ -222,11 +212,11 @@ class TestAPIv2(MailmanAPITestCase):
 
     def test_create_public_list(self):
         new_list = "new_list"
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456',
                 'archive_private': 0}
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         mlist = MailList.MailList(new_list)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(int(mlist.archive_private), 0)
@@ -235,11 +225,11 @@ class TestAPIv2(MailmanAPITestCase):
 
     def test_create_list_archive_private_out_of_max_range(self):
         new_list = "new_list"
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456',
                 'archive_private': 2}
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         mlist = MailList.MailList(new_list)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(int(mlist.archive_private), 0)
@@ -248,11 +238,11 @@ class TestAPIv2(MailmanAPITestCase):
 
     def test_create_list_archive_private_out_of_min_range(self):
         new_list = "new_list"
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456',
                 'archive_private': -1}
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         mlist = MailList.MailList(new_list)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(int(mlist.archive_private), 0)
@@ -261,11 +251,11 @@ class TestAPIv2(MailmanAPITestCase):
 
     def test_create_confirm_list(self):
         new_list = "new_list"
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456',
                 'subscribe_policy': 1}
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         mlist = MailList.MailList(new_list)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(int(mlist.subscribe_policy), 1)
@@ -274,11 +264,11 @@ class TestAPIv2(MailmanAPITestCase):
 
     def test_create_approval_list(self):
         new_list = "new_list"
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456',
                 'subscribe_policy': 2}
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         mlist = MailList.MailList(new_list)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(int(mlist.subscribe_policy), 2)
@@ -287,11 +277,11 @@ class TestAPIv2(MailmanAPITestCase):
 
     def test_create_confirm_and_approval_list(self):
         new_list = "new_list"
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456',
                 'subscribe_policy': 3}
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         mlist = MailList.MailList(new_list)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(int(mlist.subscribe_policy), 3)
@@ -300,11 +290,11 @@ class TestAPIv2(MailmanAPITestCase):
 
     def test_create_list_subscribe_policy_out_of_max_range(self):
         new_list = "new_list"
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456',
                 'subscribe_policy': 4}
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         mlist = MailList.MailList(new_list)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(int(mlist.subscribe_policy), 1)
@@ -313,11 +303,11 @@ class TestAPIv2(MailmanAPITestCase):
 
     def test_create_list_subscribe_policy_out_of_min_range(self):
         new_list = "new_list"
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456',
                 'subscribe_policy': 0}
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         mlist = MailList.MailList(new_list)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(int(mlist.subscribe_policy), 1)
@@ -326,28 +316,28 @@ class TestAPIv2(MailmanAPITestCase):
 
     def test_create_list_invalid_params(self):
         new_list = "new_list"
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': '123456',
                 'subscribe_policy': 'Invalid', 'archive_private': 'Invalid'}
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(resp.json, 14)
         self.remove_list(new_list)
 
     def test_create_list_invalid_password(self):
         new_list = "new_list"
-        url = self.url + 'lists/' + new_list
+        url = self.url + new_list
 
         data = {'admin': self.data['address'], 'password': ''}
-        resp = self.client.post(url, data, expect_errors=True)
+        resp = self.client.put(url, data, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(resp.json, 11)
         self.remove_list(new_list)
 
     def test_members(self):
         list_name = 'list13'
-        path = 'members/'
+        path = '/members'
         user_desc = UserDesc.UserDesc(self.data['address'], 'fullname', 1)
 
         self.create_list(list_name)
@@ -357,14 +347,37 @@ class TestAPIv2(MailmanAPITestCase):
         mlist.Save()
         mlist.Unlock()
 
-        resp = self.client.get(self.url + path + list_name, expect_errors=True)
+        resp = self.client.get(self.url + list_name + path, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual([self.data['address']], resp.json)
 
+    def test_member_singular(self):
+        #TODO
+        pass
+
     def test_members_unknown_list(self):
         list_name = 'list14'
-        path = 'members/'
+        path = '/members'
 
-        resp = self.client.get(self.url + path + list_name, expect_errors=True)
+        resp = self.client.get(self.url + list_name + path, expect_errors=True)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(12, resp.json)
+
+    def test_delete_list(self):
+        #TODO place holder for future addition of list deletion
+        pass
+
+    def test_list_attr(self):
+        resp = self.client.get(self.url + self.list_name, expect_errors=True)
+        total_lists = len(resp.json)
+        found = False
+
+        self.assertEqual(resp.status_code, 200)
+        self.assertIsInstance(resp.json, list)
+        self.assertEqual(total_lists, 1)
+
+        for mlist in resp.json:
+            self.assertIsInstance(mlist, dict)
+            if mlist.get("listname") == self.list_name:
+                found = True
+        self.assertTrue(found)
